requirements.txt
Flask==0.10.1
pymongo==3.0.3
beautifulsoup4==4.4.1
requests==2.8.0
passlib==1.6.5
python-dateutil=2.6.0